added allauth custom adapter
