+++ src/feedback/templatetags/translate_url.py
@@ -9,19 +9,22 @@ def translate_url(context, language):
     '''
     used to translate urls for switching languages
     '''
-    try:
-        view = resolve(context['request'].path_info)
-    except Resolver404:
+    if 'request' in context : # during some errors this prevents server error
+        try:
+            view = resolve(context['request'].path_info)
+        except Resolver404:
+            return ""
+
+        request_language = translation.get_language()
+        translation.activate(language)
+
+        namespace = view.namespace
+        view_name = f"{namespace}:{view.url_name}" if namespace else view.url_name
+        
+        url = reverse(view_name, args=view.args, kwargs=view.kwargs)
+        
+        translation.activate(request_language)
+        return url
+    else :
         return ""
-
-    request_language = translation.get_language()
-    translation.activate(language)
-
-    namespace = view.namespace
-    view_name = f"{namespace}:{view.url_name}" if namespace else view.url_name
     
-    url = reverse(view_name, args=view.args, kwargs=view.kwargs)
-    
-    translation.activate(request_language)
-    return url
-
